+++ CHANGELOG.md
@@ -15,11 +15,14 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 - Lyrics editor dialog with manual Genius search and textarea to edit the text
 - Buttons for switching between help and various lyrics options and the editor
 - Toggle for enabling or disabling Genius API requests (if disabled, folder and cache only)
+- Support for other files besides mp3, most audio files and media containers are now supported
+- Many additional fields to search for and filter on when queueing songs
 
 ### Changed
 
 - Fullscreen layout styling related to the "up next" notice and general margins
 - Layout of the status line and selection colors
+- Show an event for each failed file so you know that to do about it
 
 ### Fixed